backend/usiu_smart_nav/requirements.txt
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-leaflet==0.32.0
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework-gis==1.2.0
idna==3.10
psycopg2-binary==2.9.10
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
networkx==3.3